refactor (kitchen pos) : “ProductDao 인터페이스 추출 및 기존 ProductDao 클래스 이름을 JdbcTemplateProductDao로 변경
